Segmented main_entities.py into separate classes.
