+++ Main Entities/BattleSimulation.py
@@ -1,0 +1,7 @@
+from Trainer import Trainer
+
+class BattleSimulation:
+    def __init__(self, trainer1: Trainer, trainer2: Trainer, trainer1AI: bool,
+                 trainer2AI: bool, trainer1AIPolicy, trainer2AIPolicy) -> None:
+        pass
+
