+++ Main Entities/Team.py
@@ -1,0 +1,3 @@
+class Team:
+    def __init__(self, all_pokemon):
+        pass
