+++ Main Entities/Trainer.py
@@ -1,0 +1,3 @@
+class Trainer:
+    pass
+
